fix all eslint errors

+++ src/components/ItemInputCard.tsx
@@ -1,20 +1,25 @@
 import { AvatarSplittingButton } from '@/components/Avatar';
 import { DeleteButton } from '@/components/DeleteButton';
 import { Input } from '@/components/Input';
-import { usePeopleStore } from '@/hooks/usePeopleStore';
+import { IntegerInput } from '@/components/IntegerInput';
 import { useItemsStore } from '@/hooks/useItemsStore';
+import { usePeopleStore } from '@/hooks/usePeopleStore';
 import { currencyInput } from '@/utils/currencyInput';
 import { formatCurrency } from '@/utils/formatCurrency';
-import { IntegerInput } from './IntegerInput';
 
 export function ItemInputCard({ itemId }: { itemId: string }) {
-  const item = useItemsStore((state) => state.items.get(itemId))!;
-  const { changeItemProp, deleteItem, shareItem } = useItemsStore((state) => ({
+  const item = useItemsStore(state => state.items.get(itemId));
+  if (!item) {
+    console.log('an error that should never happen');
+    return null;
+  }
+
+  const { changeItemProp, deleteItem, shareItem } = useItemsStore(state => ({
     changeItemProp: state.changeItemProp,
     deleteItem: state.deleteItem,
     shareItem: state.shareItem,
   }));
-  const { people } = usePeopleStore((state) => ({
+  const { people } = usePeopleStore(state => ({
     people: state.people,
   }));
 
@@ -27,24 +32,23 @@ export function ItemInputCard({ itemId }: { itemId: string }) {
             className="w-full"
             placeholder={item.defaultName}
             value={item.name}
-            onChange={(value) => changeItemProp(item, 'name', value.toUpperCase())}
+            onChange={value =>
+              changeItemProp(item, 'name', value.toUpperCase())
+            }
           />
           <DeleteButton onClick={() => deleteItem(item.id)} />
         </div>
         <div>
-          Preço unitário: R$
-          {' '}
+          Preço unitário: R${' '}
           <Input
             className="w-20 text-right"
             maxLength={11}
             inputMode="numeric"
             placeholder="0,00"
             value={item.unitPrice}
-            onChange={(value) => changeItemProp(
-              item,
-              'unitPrice',
-              currencyInput.toCents(value),
-            )}
+            onChange={value =>
+              changeItemProp(item, 'unitPrice', currencyInput.toCents(value))
+            }
             format={currencyInput.format}
           />
         </div>
@@ -52,8 +56,10 @@ export function ItemInputCard({ itemId }: { itemId: string }) {
           <span>Quantidade: </span>
           <IntegerInput
             value={item.quantity}
-            onChange={(event) => changeItemProp(item, 'quantity', Number(event.target.value))}
-            buttonsFunction={(newValue) => changeItemProp(item, 'quantity', newValue)}
+            onChange={value => changeItemProp(item, 'quantity', Number(value))}
+            buttonsFunction={newValue =>
+              changeItemProp(item, 'quantity', newValue)
+            }
           />
         </div>
         <div className="whitespace-nowrap">
@@ -62,7 +68,7 @@ export function ItemInputCard({ itemId }: { itemId: string }) {
         </div>
       </div>
       <div className="py-2 flex flex-wrap gap-y-2 justify-around">
-        {[...people.values()].map((person) => (
+        {[...people.values()].map(person => (
           <div
             className="flex flex-col items-center gap-y-1 w-[4.6rem]"
             key={person.id}
@@ -80,10 +86,12 @@ export function ItemInputCard({ itemId }: { itemId: string }) {
                 className="w-full text-center px-0"
                 type="text"
                 value={item.sharedBy[person.id] as number}
-                onChange={(value) => changeItemProp(item, 'sharedBy', {
-                  ...item.sharedBy,
-                  [person.id]: currencyInput.toCents(value),
-                })}
+                onChange={value =>
+                  changeItemProp(item, 'sharedBy', {
+                    ...item.sharedBy,
+                    [person.id]: currencyInput.toCents(value),
+                  })
+                }
                 format={currencyInput.format}
                 inputMode="numeric"
                 placeholder="0,00"
